refactor: move multer js from utils to middlewares folder
